GCC warnings fixes

+++ src/engines/synths/OTTOFM/ottofm.cpp
@@ -589,9 +589,9 @@ namespace otto::engines {
       ctx.textAlign(HorizontalAlign::Left, VerticalAlign::Middle);
       ctx.fillText("Self-mod", {x_pad_left, y_pad + 100});
 
-      int circ_x = 2 * x_pad_left;
-      int circ_y = 2.2 * y_pad;
-      int rad = 15;
+      float circ_x = 2 * x_pad_left;
+      float circ_y = 2.2 * y_pad;
+      float rad = 15;
 
 
       ctx.beginPath();
